[REFACTOR] `can_run_services_as_svc_user` moved to process module
Taking a cue from the Linux implementation, this function is really
based on the process itself, rather than anything dealing with users.
